refactor(v2): delete unused epic-* scripts under .agents/scripts
Remove the unwired Epic CLIs (audit-prepare/recheck, plan-clarity,
plan-healthcheck) and the dead epic-plan-state-store / scoreEpicBody
surfaces. Extract live nav helpers into plan-navigation.js and move
resolveAuditLenses into audit-lens-routing.js. Drop plan-context --epic
mode and Epic-close helpers from code-review. Refresh the maintainability
snapshot for the intentional MI floor move on code-review.js and removed
CLI rows.

baseline-refresh: true
