+++ .agents/scripts/lib/config-settings-schema.js
@@ -297,8 +297,7 @@ const PLANNING_SCHEMA = {
     failOnLargeFanOut: { type: 'boolean' },
     largeFanOutThreshold: { type: 'integer', minimum: 0 },
     // Navigability-reachability config consumed by the plan-persist draft
-    // reachability gate (Epic #4131 F7; demoted into persist by #4474 PR6 —
-    // the manual epic-plan-healthcheck --paranoid re-check reads it too).
+    // reachability gate (Epic #4131 F7; demoted into persist by #4474 PR6).
     // Opt-in: absent or empty routeGlobs degrades to a silent no-op.
     navigation: {
       type: 'object',
