+++ .agents/scripts/lib/orchestration/code-review.js
@@ -36,11 +36,8 @@
  */
 
 import { hasSurvivingCritical } from '../audit-suite/findings.js';
-import { resolveLensTier } from '../audit-suite/selector.js';
 import { resolveConfig } from '../config-resolver.js';
-import { selectAuditStrategy } from '../dynamic-workflow/capability.js';
 import { gitSpawn } from '../git-utils.js';
-import { read as readPlanState } from './epic-plan-state-store.js';
 import { resolveDepth } from './review-depth.js';
 import {
   countBySeverity,
@@ -91,230 +88,6 @@ function countChangedFiles({ baseRef, headRef, gitSpawnFn = gitSpawn }) {
   } catch {
     return null;
   }
-}
-
-/**
- * Producer-side resolver (Story #3937, extended by Story #3938): read an
- * Epic's judged `planningRisk` envelope off its `epic-plan-state` checkpoint
- * and resolve the review depth via the shared {@link resolveDepth} resolver.
- * This is the live epic-scope counterpart to `resolveRiskRoutedLenses` in
- * `epic-audit-prepare.js` — it reuses the same best-effort `readPlanState`
- * pattern so the producer never aborts the review on a risk-read failure.
- *
- * Best-effort and total: a missing/unparseable checkpoint, an absent
- * `planningRisk` field, a read failure, or a malformed `overallLevel` all
- * degrade to `standard` (never throws). So an Epic that skipped `/plan`
- * (no checkpoint) still gets a passing `standard` review with no new failure
- * mode. The producer resolves depth from the judged risk alone (the diff is
- * not yet enumerated at this point); the full risk + diff-width combination is
- * applied in `runCodeReview`, which knows the changed-file count. A high-risk
- * Epic resolves to `deep` here; a low-risk one to `light` (the diff width can
- * only escalate, never downgrade, that producer-side resolution).
- *
- * @param {{
- *   epicId: number,
- *   provider: object,
- *   readPlanState?: typeof readPlanState,
- * }} params
- * @returns {Promise<ReviewDepth>}
- */
-export async function resolveReviewDepthForEpic({
-  epicId,
-  provider,
-  readPlanState: readPlanStateFn = readPlanState,
-}) {
-  let state = null;
-  try {
-    state = await readPlanStateFn({ provider, epicId });
-  } catch {
-    // A read failure (provider error, malformed comment) must not abort the
-    // review. Degrade to the neutral `standard` depth.
-    return 'standard';
-  }
-  return resolveDepth({ overallLevel: state?.planningRisk?.overallLevel });
-}
-
-/**
- * The axes whose presence (at `high` risk) routes a specific post-delivery
- * audit lens. Mirrors the audit-workflow names under
- * `.agents/workflows/audit-*.md`. Story #3939 broadened the routing so every
- * high-risk REQUIRED axis maps to a lens rather than leaving three of them
- * (`data-migration`, `destructive-mutation`, `billing`) routing nothing:
- *
- *   - `security`             → `audit-security`     (auth/secret boundary)
- *   - `public-api`           → `audit-architecture` (the canonical
- *                              architectural axis; a breaking public surface)
- *   - `data-migration`       → `audit-quality`      (migration correctness +
- *                              regression coverage)
- *   - `destructive-mutation` → `audit-security`     (irreversible mutation is
- *                              an auth/abuse boundary; co-routes with
- *                              `security`, de-duplicated)
- *   - `billing`              → `audit-privacy`      (money + the PII/consent
- *                              surface that travels with it)
- *   - `critical-workflow`    → `audit-quality`      (the load-bearing path
- *                              warrants the deepest coverage pass)
- *
- * The `visible-behavior` axis intentionally routes NO lens: it forces the
- * acceptance-spec disposition at plan time and no audit lens maps cleanly to
- * it. Any other axis (or a low/medium-risk axis) contributes no lens.
- *
- * Every key here MUST be a value in the `axis` enum of
- * `.agents/schemas/risk-verdict.schema.json` — the verdict-derived envelope
- * can only ever carry schema-valid axes, so a key absent from that enum is
- * unreachable dead routing (Story #3889 removed the unreachable
- * `architecture` key; the architectural axis is `public-api`).
- */
-const AXIS_TO_LENS = Object.freeze({
-  security: 'audit-security',
-  'public-api': 'audit-architecture',
-  'data-migration': 'audit-quality',
-  'destructive-mutation': 'audit-security',
-  billing: 'audit-privacy',
-  'critical-workflow': 'audit-quality',
-});
-
-/**
- * Stable output order for routed lenses so the lens list is deterministic
- * regardless of axis ordering in the verdict. Every lens any axis can route
- * appears here exactly once; `resolveAuditLenses` filters this canonical
- * order down to the matched set (Story #3939).
- */
-const LENS_ORDER = Object.freeze([
-  'audit-security',
-  'audit-architecture',
-  'audit-quality',
-  'audit-privacy',
-]);
-
-/**
- * Resolve the set of post-delivery audit lenses a judged risk envelope routes.
- *
- * High-risk axes map to their audit lens via {@link AXIS_TO_LENS}; only axes
- * judged `high` contribute (a `low`/`medium` axis carries no lens). The result
- * is de-duplicated and stably ordered by {@link LENS_ORDER} so an envelope
- * listing the `public-api` axis more than once routes `['audit-architecture']`
- * once, not twice — and two distinct axes routing the same lens (e.g.
- * `security` + `destructive-mutation` both → `audit-security`) collapse to a
- * single entry. A low-risk envelope — or any envelope with no high-risk routed
- * axis — resolves to an empty array (no lens beyond the existing baseline
- * gates).
- *
- * Pure function — no I/O, no side effects.
- *
- * @param {{ axes?: Array<{ axis?: string, level?: string }> }} [envelope]
- * @returns {string[]} Ordered, de-duplicated audit-lens identifiers.
- */
-export function resolveAuditLenses(envelope = {}) {
-  const axes = Array.isArray(envelope?.axes) ? envelope.axes : [];
-  const matched = new Set();
-  for (const entry of axes) {
-    if (!entry || entry.level !== 'high') continue;
-    const lens = AXIS_TO_LENS[entry.axis];
-    if (lens) matched.add(lens);
-  }
-  return LENS_ORDER.filter((lens) => matched.has(lens));
-}
-
-/**
- * Compose the Epic-close (gate3) lens roster from the change-set selection
- * plus every risk-routed lens.
- *
- * STOPGAP (post-#4405 review finding): Story #4412 excluded every
- * `local`-tier change-set lens here on the premise that the Story-scope
- * local-lens pass (Story #4409) verifies those concerns shift-left. That
- * premise does not hold yet — `runLocalLensReview`
- * (story-close/phases/code-review.js) materializes lens bodies with no
- * `artifactPrefix` (so no content persists) and no consumer acts on its
- * result, which left change-set lenses (security, privacy, performance,
- * quality, …) executed by NO reviewer at ANY tier. Local-tier lenses are
- * therefore KEPT at Epic close until the story-scope pass has a real
- * consumer; re-enable the exclusion then (tracked as a follow-up issue).
- * The write-time maker checklists (#4410) are advisory self-checks, not
- * verification. The `resolveLensTierFn` seam is retained so the
- * re-exclusion is a one-line revert.
- *
- * The two inputs mirror the `epic-audit-prepare.js` envelope:
- *   - `changeSetAudits` — the raw change-set gate3 selection, kept in full
- *     (see the stopgap note above).
- *   - `riskRoutedAudits` — the verdict-routed high-risk lenses plus the
- *     route-glob navigability lens. Kept in **full**, regardless of tier: a
- *     high-risk axis (or a route-adding change set) that demands a local-tier
- *     lens still runs it at Epic close, because that demand is the whole point
- *     of risk routing.
- *
- * De-duplicated and order-preserving (kept change-set lenses first, then the
- * risk-routed extras). Pure over the injected `resolveLensTierFn` seam; the
- * default resolver reads `audit-rules.json` from disk.
- *
- * @param {{
- *   changeSetAudits?: string[],
- *   riskRoutedAudits?: string[],
- *   resolveLensTierFn?: typeof resolveLensTier,
- * }} [params]
- * @returns {string[]} The Epic-close roster (all change-set tiers +
- *   risk-routed; local-tier exclusion suspended per the stopgap above).
- */
-export function selectEpicCloseLenses({
-  changeSetAudits = [],
-  riskRoutedAudits = [],
-  resolveLensTierFn = resolveLensTier,
-} = {}) {
-  const kept = [];
-  const seen = new Set();
-  const add = (lens) => {
-    if (typeof lens !== 'string' || lens.length === 0 || seen.has(lens)) return;
-    seen.add(lens);
-    kept.push(lens);
-  };
-  for (const lens of Array.isArray(changeSetAudits) ? changeSetAudits : []) {
-    if (typeof lens !== 'string' || lens.length === 0) continue;
-    // Stopgap: keep local-tier lenses (see docstring). The tier seam is
-    // deliberately untouched so restoring `=== 'local' → continue` is a
-    // one-line revert once the story-scope tier has a real consumer.
-    void resolveLensTierFn;
-    add(lens);
-  }
-  for (const lens of Array.isArray(riskRoutedAudits) ? riskRoutedAudits : []) {
-    add(lens);
-  }
-  return kept;
-}
-
-/**
- * Build the post-delivery audit-lens execution plan for a judged risk
- * envelope. Each routed lens (see {@link resolveAuditLenses}) is paired with a
- * strategy decision from the **existing** `selectAuditStrategy` engine — no new
- * audit machinery is introduced. A low-risk envelope resolves to an empty
- * `lenses` array and runs no audit beyond the baseline gates (Story #3876).
- *
- * Pure with respect to the injected `selectAuditStrategyFn` (default is the
- * shared dynamic-workflow engine, which is itself pure over its snapshot).
- *
- * @param {{ axes?: Array<{ axis?: string, level?: string }> }} [envelope]
- * @param {{
- *   snapshot?: object,
- *   forceStrategy?: ('orchestrated'|'sequential'|null),
- *   selectAuditStrategyFn?: typeof selectAuditStrategy,
- * }} [opts]
- * @returns {{ lenses: string[], plan: Array<{ lens: string, strategy: string, reason: string, forced: boolean }> }}
- */
-export function planAuditLenses(envelope = {}, opts = {}) {
-  const {
-    snapshot = {},
-    forceStrategy = null,
-    selectAuditStrategyFn = selectAuditStrategy,
-  } = opts;
-  const lenses = resolveAuditLenses(envelope);
-  const plan = lenses.map((lens) => {
-    const decision = selectAuditStrategyFn({ snapshot, forceStrategy });
-    return {
-      lens,
-      strategy: decision.strategy,
-      reason: decision.reason,
-      forced: decision.forced,
-    };
-  });
-  return { lenses, plan };
 }
 
 /**
