+++ .agents/scripts/lib/orchestration/ticketing/reads.js
@@ -103,9 +103,8 @@ export const STRUCTURED_COMMENT_TYPES = Object.freeze([
   // surface (Epic #1030).
   'story-perf-summary',
   'epic-perf-report',
-  // Story #2128 — Phase 6 Epic Clarity Gate. `epic-plan-clarity.js` upserts
-  // a `clarity-gate-update` comment on the Epic when the operator approves
-  // a sharpened body rewrite, recording the persistence event for audit.
+  // Story #2128 — Phase 6 Epic Clarity Gate (CLI retired). Historical
+  // `clarity-gate-update` comments may still exist on older tickets.
   'clarity-gate-update',
   // Story #2635 — Tech Spec freshness check. `plan-persist.js`
   // upserts a `spec-freshness` comment on the Epic listing any
