+++ docs/data-dictionary.md
@@ -250,7 +250,7 @@ the Epic is the SSOT; the on-disk file is a renderer cache regenerable via
 | `signals-writer.appendSignal`                       | Helper   | Append-only NDJSON writer at `lib/observability/signals-writer.js`. Writes one JSON record per line to `temp/epic-<eid>/stories/story-<sid>/signals.ndjson`. Consumers: `diagnose-friction.js`, `story-close.js` reap-failure (via `post-merge-pipeline.js`), `analyze-execution.js`, and the retro signal gatherer (`lib/orchestration/retro/phases/gather-signals.js`). Per-kind quality-gate logic formerly in `check-crap.js` / `check-maintainability.js` now lives in `lib/baselines/kinds/{lint,coverage,crap,maintainability,mutation}.js` behind `check-baselines.js`. Replaced the deleted in-process emitter class in Epic #1030 Story #1042. |
 | `--reap-discard-after-merge` / `--no-reap-discard-after-merge` | CLI flag | `/deliver` Phase 7 flag. Default force-reaps worktrees whose Story branch is already merged into `epic/<id>` (per `git merge-base --is-ancestor`), discarding uncommitted post-merge drift; the `--no-` form preserves prior skip-on-uncommitted behavior. Force-reap emits a `friction` comment listing discarded paths. |
 | Version-bump-intent snapshot                        | Checkpoint | `/deliver` Phase 0.5 parses the Epic body for `Release target:` / `--segment` directives and posts a `notification` structured comment on the Epic (marker `<!-- notification: version-bump-intent -->`) when they disagree with `release.autoVersionBump`.            |
-| Launcher-level config validation                    | Contract | `validateOrchestrationConfig(config)` (from `lib/config/validate-orchestration.js`) runs at launcher startup in `plan-context.js`, `plan-persist.js`, `epic-plan-clarity.js`, `epic-plan-healthcheck.js`, `bootstrap.js`, and `agents-bootstrap-github.js` — a schema-invalid `.agentrc.json` exits non-zero before any long-running flow begins. |
+| Launcher-level config validation                    | Contract | `validateOrchestrationConfig(config)` (from `lib/config/validate-orchestration.js`) runs at launcher startup in `plan-context.js`, `plan-persist.js`, `bootstrap.js`, and `agents-bootstrap-github.js` — a schema-invalid `.agentrc.json` exits non-zero before any long-running flow begins. |
 
 ---
